Create Lab9_#1
Pascal’s triangle, in algebra, a triangular arrangement of numbers that gives the coefficients in the expansion of any binomial expression, such as (x + y)^n.

For a given integer N, print the first N rows of Pascal's Triangle. Print each row with each value separated by a single space. The value at the nth row and rth column of the triangle is equal to n!/(r!*(n-r)!) where indexing starts from 0.