Create Lab10_#3
A number is called a smart number if it has an odd number of factors. Given some numbers, find whether they are smart numbers or not. Write a function check_smart(int n) for checking that given number is smart number or not and call it in driver code for 't' test cases.